rename: change layout file name
