+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App.tsx";
-import Layout from "./layout/Layout.tsx";
+import Layout from "./layout/index.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
